+++ subot/settings.py
@@ -41,6 +41,7 @@ class Config:
     master_volume: int = 100
     main_volume: int = 100
     altar: int = 100
+    blacksmith: int = 100
     chest: int = 100
     emblem: int = 100
     exotic_portal: int = 100
@@ -85,6 +86,7 @@ class Config:
         ini["VOLUME"] = {
             "main_volume": self.main_volume,
             "altar": self.altar,
+            "blacksmith": self.blacksmith,
             "chest": self.chest,
             "emblem": self.emblem,
             "exotic_portal": self.exotic_portal,
@@ -127,6 +129,7 @@ class Config:
 
         default_config.master_volume = volume.getint("main_volume", fallback=default_config.master_volume)
         default_config.altar = volume.getint("altar", fallback=default_config.altar)
+        default_config.blacksmith = volume.getint("blacksmith", fallback=default_config.blacksmith)
         default_config.chest = volume.getint("chest", fallback=default_config.chest)
         default_config.emblem = volume.getint("emblem", fallback=default_config.emblem)
         default_config.exotic_portal = volume.getint("exotic_portal", fallback=default_config.exotic_portal)